fix: change type of minDistance

+++ core-dto/src/main/java/com/stride/tracking/core/dto/route/request/RouteFilter.java
@@ -13,5 +13,5 @@ public class RouteFilter {
     private String userId;
     private String sportId;
     private Boolean publicRoute;
-    private Integer minDistance;
+    private Double minDistance;
 }
